Manual
manual. 
now the program works without using UME. the program has 24 script files, all of them have ! at the beginning of name. on the left, the file name indicates the number of frames in the video cassette that will be created. on the right, the script file name indicates compatibility with the minecraft version. double-click on the script file that you need. the console will open and after a while the inscription will appear (enter the starting map uuid and click enter button) enter the starting uuid and press enter. at the end of the process, a file will appear in the folder !!out.schematic the rest of the instructions on how to transfer videos to the world of bedrock can be seen here.